Fix for 'Blown Fuse' Issue
It was discovered that the current compensation control loop in firmware had a bug where under certain conditions, the batlab would overshoot its setpoint target at the start of a discharge test and blow the fuses.

The fix implemented here is turning off the control loop in firmware (not allowing you to turn it on) and then implementing the control loop in software instead of firmware.