Added Constant in REG_SETTINGS for firmware version 5
Setting SET_NO_PSU_DCHG_ENABLE allows you to bypass the rue created in firmware version 5 wherein discharging is only allowed if the external power supply is plugged in. Newer Batlabs put the fan on the EXT PWR line to put less strain on USB, which creates a requirement for have EXT PWR plugged in for discharge so that the fan can run. But if you have an older batlab or Mod your Batlab so that the fan is back on USB power, then you will want to enable this settings so you can use discharge mode wthout plugging in a supply