Added contast voltage charge/discharge feature
Works best with cell impedances associated with standard 2-3 AH 18650 cells.

To enable constant voltage mode, set "constantVoltageEnable" attribute to true in the settings JSON file.  THis is meant to be used in conjuction with a cycletest or dischargetest  - it works by gradually scaling down the current when the cell is approching its voltage limits in either the charge or discharge direction. The algorithm is such that it gradually settles on the desired voltage limit in a smooth manner rather than chainging from CC to CV step in a piecewise manner.